Allow certain functions with Policies - Video 19

+++ app/Http/Controllers/Admin/PostController.php
@@ -71,6 +71,9 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+
+        $this->authorize('author', $post);
+
         $categories = Category::pluck('name', 'id');
         $tags = Tag::all();
 
@@ -82,6 +85,9 @@ class PostController extends Controller
      */
     public function update(PostRequest $request, Post $post)
     {
+
+        $this->authorize('author', $post);
+
         $post->update($request->all());
 
         if ($request->file('file')) {
@@ -116,6 +122,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+
+        $this->authorize('author', $post);
+
         $post->delete();
 
         return redirect()->route('admin.posts.index', $post)->with('info', 'El post se eliminó con éxito');
